+++ src/entities/auth/api/auth.ts
@@ -25,12 +25,3 @@ export async function logout(accessToken: string | null) {
     headers: { Authorization: `Bearer ${accessToken}` },
   });
 }
-
-export async function withdrawalAccount(accessToken: string | null) {
-  if (!accessToken) throw new Error('로그인을 해주세요.');
-
-  await fetchAPI('/login/account', {
-    method: 'DELETE',
-    headers: { Authorization: `Bearer ${accessToken}` },
-  });
-}
